+++ src/ResourceManager/Resources/Commands.ResourceManager/Cmdlets/Implementation/ResourceGroupDeployments/NewAzureResourceGroupDeploymentCmdlet.cs
@@ -67,7 +67,7 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation
                 ResourceGroupName,
                 () =>
                 {
-                    PSCreateResourceGroupDeploymentParameters parameters = new PSCreateResourceGroupDeploymentParameters()
+                    PSDeploymentCmdletParameters parameters = new PSDeploymentCmdletParameters()
                     {
                         ResourceGroupName = ResourceGroupName,
                         DeploymentName = Name,